wait time is long

+++ src/de/haw/rnp01/newsticker/model/RandomGenerator.java
@@ -44,7 +44,7 @@ public class RandomGenerator {
      *
      * @return sleep time from 1000 - 5000 ms
      */
-    public int generateRandomSleepTime() {
+    public long generateRandomSleepTime() {
         return (this.generator.nextInt(5000) + 1000);
     }
